# job-platform/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MedMatch Job Platform API",
    "description": "API for job matching platform connecting medical professionals with healthcare employers. Includes application tracking, dashboard analytics, and notification system.",
    "version": "1.1.0",
    "contact": {
      "name": "MedMatch API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/applications": {
      "get": {
        "summary": "List applications",
        "description": "Get all applications with filters and pagination",
        "tags": ["Applications"],
        "parameters": [
          {
            "name": "candidateId",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "Filter by candidate ID"
          },
          {
            "name": "jobId",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "Filter by job ID"
          },
          {
            "name": "employerId",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "Filter by employer ID"
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["pending", "reviewed", "accepted", "rejected"] },
            "description": "Filter by application status"
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20 },
            "description": "Items per page (max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of applications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Application" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit application",
        "description": "Submit a new job application",
        "tags": ["Applications"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["candidateId", "jobId"],
                "properties": {
                  "candidateId": { "type": "integer", "description": "ID of the candidate" },
                  "jobId": { "type": "integer", "description": "ID of the job" },
                  "coverLetter": { "type": "string", "description": "Optional cover letter" },
                  "cvUrl": { "type": "string", "description": "URL to candidate CV" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/Application" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed" },
          "404": { "description": "Candidate or job not found" },
          "409": { "description": "Already applied to this job" }
        }
      }
    },
    "/applications/{id}": {
      "get": {
        "summary": "Get application details",
        "description": "Get details of a specific application",
        "tags": ["Applications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Application ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Application details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/ApplicationDetail" }
                  }
                }
              }
            }
          },
          "404": { "description": "Application not found" }
        }
      },
      "patch": {
        "summary": "Update application status",
        "description": "Update the status of an application (employer only)",
        "tags": ["Applications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Application ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { 
                    "type": "string", 
                    "enum": ["pending", "reviewed", "accepted", "rejected"],
                    "description": "New application status"
                  },
                  "notes": { "type": "string", "description": "Employer notes" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Application updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/Application" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid status" },
          "404": { "description": "Application not found" }
        }
      },
      "delete": {
        "summary": "Withdraw application",
        "description": "Soft delete an application (candidate can withdraw)",
        "tags": ["Applications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Application ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Application withdrawn successfully"
          },
          "404": { "description": "Application not found" }
        }
      }
    },
    "/candidates/{id}/applications": {
      "get": {
        "summary": "Get candidate applications",
        "description": "Get all applications for a specific candidate",
        "tags": ["Candidates", "Applications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Candidate ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of candidate applications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "candidateId": { "type": "integer" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Application" }
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "byStatus": {
                          "type": "object",
                          "properties": {
                            "pending": { "type": "integer" },
                            "reviewed": { "type": "integer" },
                            "accepted": { "type": "integer" },
                            "rejected": { "type": "integer" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Candidate not found" }
        }
      }
    },
    "/jobs/{id}/applications": {
      "get": {
        "summary": "Get job applications",
        "description": "Get all applications for a specific job (employer only)",
        "tags": ["Jobs", "Applications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of job applications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobId": { "type": "integer" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Application" }
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "byStatus": {
                          "type": "object",
                          "properties": {
                            "pending": { "type": "integer" },
                            "reviewed": { "type": "integer" },
                            "accepted": { "type": "integer" },
                            "rejected": { "type": "integer" }
                          }
                        },
                        "averageMatchScore": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Job not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Application": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "candidateId": { "type": "integer" },
          "jobId": { "type": "integer" },
          "status": { "type": "string", "enum": ["pending", "reviewed", "accepted", "rejected"] },
          "coverLetter": { "type": "string" },
          "cvUrl": { "type": "string" },
          "matchScore": { "type": "integer" },
          "appliedAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "candidateFirstName": { "type": "string" },
          "candidateLastName": { "type": "string" },
          "candidateEmail": { "type": "string" },
          "candidateSpecialty": { "type": "string" },
          "jobTitle": { "type": "string" },
          "jobSpecialty": { "type": "string" },
          "jobLocation": { "type": "string" },
          "employerName": { "type": "string" },
          "employerId": { "type": "integer" }
        }
      },
      "ApplicationDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Application" },
          {
            "type": "object",
            "properties": {
              "jobDescription": { "type": "string" },
              "jobRequirements": { "type": "string" },
              "salaryMin": { "type": "integer" },
              "salaryMax": { "type": "integer" },
              "jobType": { "type": "string" },
              "experienceRequired": { "type": "integer" },
              "candidateExperience": { "type": "integer" },
              "employerType": { "type": "string" }
            }
          }
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "hasNextPage": { "type": "boolean" },
          "hasPrevPage": { "type": "boolean" }
        }
      }
    }
  }
}